Added note for using multiple endpoints
Need to clarify that to be able to register a resource with multiple private DNS zones is done by creating multiple private-endpoints and registering each to the relevant private dns zone to facilitate resolution and traffic segregation. As currently the bicep code for private-endpoints does not allow registration to multiple private dns zones.